Complete transition to plugins
