+++ plugins/cohere/superduper_cohere/__init__.py
@@ -1,7 +1,3 @@
-from superduper.misc.annotations import requires_packages
-
-_, requirements = requires_packages(['cohere', '4.40'])
-
 from .model import CohereEmbed, CohereGenerate
 
 __version__ = "0.0.5"
