+++ plugins/mongodb/superduper_mongodb/__init__.py
@@ -2,6 +2,7 @@ from .artifacts import MongoArtifactStore as ArtifactStore
 from .data_backend import MongoDataBackend as DataBackend
 from .metadata import MongoMetaDataStore as MetaDataStore
 from .query import MongoQuery
+from .vector_search import MongoAtlasVectorSearcher as VectorSearcher
 
 __version__ = "0.0.5"
 
@@ -10,4 +11,5 @@ __all__ = [
     "MongoQuery",
     "DataBackend",
     "MetaDataStore",
+    "VectorSearcher",
 ]
