+++ plugins/openai/superduper_openai/__init__.py
@@ -1,7 +1,3 @@
-from superduper.misc.annotations import requires_packages
-
-_, requirements = requires_packages(['openai', '1.1.2', None], ['httpx'])
-
 from .model import OpenAIChatCompletion, OpenAIEmbedding
 
 __version__ = "0.0.6"
