+++ plugins/pillow/superduper_pillow/__init__.py
@@ -1,7 +1,3 @@
-from superduper.misc.annotations import requires_packages
-
-_, requirements = requires_packages(['PIL', '10.2.0', None, 'pillow'])
-
 from .encoder import pil_image
 
 __version__ = "0.0.5"
