+++ plugins/sentence_transformers/superduper_sentence_transformers/__init__.py
@@ -1,9 +1,5 @@
-from superduper.misc.annotations import requires_packages
-
 from .model import SentenceTransformer
 
 __version__ = "0.0.5"
 
 __all__ = ('SentenceTransformer',)
-
-_, requirements = requires_packages(['sentence-transformers', '2.2.2', None])
