+++ plugins/torch/superduper_torch/__init__.py
@@ -1,11 +1,6 @@
-from superduper.misc.annotations import requires_packages
-
-from .training import TorchTrainer
-
-_, requirements = requires_packages(['torch', '2.0.0'])
-
 from .encoder import tensor
 from .model import TorchModel, torchmodel
+from .training import TorchTrainer
 
 __version__ = "0.0.5"
