--- superduper/backends/ibis/__init__.py
@@ -1,5 +1,0 @@
-from superduper_ibis import *  # noqa
-
-from superduper.misc.annotations import warn_plugin_deprecated
-
-warn_plugin_deprecated('ibis')
