--- superduper/backends/ibis/data_backend.py
@@ -1,1 +1,0 @@
-from superduper_ibis.data_backend import *  # noqa
