--- superduper/backends/ibis/field_types.py
@@ -1,1 +1,0 @@
-from superduper_ibis.field_types import *  # noqa
