--- superduper/backends/ibis/query.py
@@ -1,1 +1,0 @@
-from superduper_ibis.query import *  # noqa
