--- superduper/backends/mongodb/__init__.py
@@ -1,5 +1,0 @@
-from superduper_mongodb import *  # noqa
-
-from superduper.misc.annotations import warn_plugin_deprecated
-
-warn_plugin_deprecated('mongodb')
