--- superduper/backends/mongodb/data_backend.py
@@ -1,1 +1,0 @@
-from superduper_mongodb.data_backend import *  # noqa
