--- superduper/backends/mongodb/query.py
@@ -1,1 +1,0 @@
-from superduper_mongodb.query import *  # noqa
