--- superduper/backends/sqlalchemy/__init__.py
@@ -1,5 +1,0 @@
-from superduper_sqlalchemy import *  # noqa
-
-from superduper.misc.annotations import warn_plugin_deprecated
-
-warn_plugin_deprecated('sqlalchemy')
