--- superduper/backends/sqlalchemy/db_helper.py
@@ -1,1 +1,0 @@
-from superduper_sqlalchemy.db_helper import *  # noqa
