--- superduper/backends/sqlalchemy/metadata.py
@@ -1,1 +1,0 @@
-from superduper_sqlalchemy.metadata import *  # noqa
