+++ superduper/cli/main.py
@@ -33,9 +33,9 @@ def start(
     :param host: Host to connect to remotely
     :param headless: Toggle to ``True`` to suppress the browser.
     """
+    from superduper import CFG
     from superduper.rest.base import SuperDuperApp
     from superduper.rest.build import build_frontend, build_rest_app
-    from superduper import CFG
 
     CFG.log_colorize = False
